refactor(data-cache): collapse the cache schema to its mobile reality
The cache schema file carried ~120 lines of desktop-ported convention
essays and a casual-key type guard (UseCacheCasualKey) that nothing in
either CacheService signature ever consumed. Rewrite the file down to
the three keys mobile actually defines — the frontend memory/persist
probes and the backend provider key-rotation cursor — keeping the
template-key type machinery that types the two live template entries.
Persisted MMKV key strings are untouched.

deepEqual moves out of the cache module to src/shared/utils: it is a
generic JSON-value comparator, and ModelService was importing the cache
module solely to reach it. Its tests move with it.
