+++ src/shared/data/cache/cacheSchemas.ts
@@ -1,116 +1,38 @@
 /**
- * Cache Schema Definitions
+ * Mobile-owned cache key tables and the type machinery that keeps them safe.
  *
- * Ported from desktop `src/shared/data/cache/cacheSchemas.ts`. Mobile has no
- * multi-window shared tier, so desktop `SharedCacheSchema` entries that mobile
- * needs get folded into {@link UseCacheSchema} (memory tier) instead.
+ * Keys follow `namespace.sub.key_name` — lowercase segments joined by dots,
+ * with `${variable}` placeholders as literal segments in template keys
+ * (`/^[a-z][a-z0-9_]*(\.[a-z][a-z0-9_]*)+$/`). Persisted MMKV key strings are
+ * data on real devices: renaming a persist key strands its stored value.
  *
- * ## Key Naming Convention
- *
- * All cache keys (fixed and template) MUST follow the format: `namespace.sub.key_name`
- *
- * Rules:
- * - At least 2 segments separated by dots (.)
- * - Each segment uses lowercase letters, numbers, and underscores only
- * - Pattern: /^[a-z][a-z0-9_]*(\.[a-z][a-z0-9_]*)+$/
- * - Template placeholders `${xxx}` are treated as literal string segments
- *
- * Examples:
- * - 'app.path.resources' (valid)
- * - 'chat.multi_select_mode' (valid)
- * - 'scroll.position.${topicId}' (valid template key)
- * - 'userAvatar' (invalid - missing dot separator)
- * - 'App.user' (invalid - uppercase not allowed)
- * - 'scroll.position:${id}' (invalid - colon not allowed)
- *
- * Desktop enforces this via the `data-schema-key/valid-key` ESLint rule; mobile
- * has no such rule yet, so the convention is upheld by review.
- *
- * ## Template Key Support
- *
- * Template keys allow type-safe dynamic keys using template literal syntax.
- * Define in schema with `${variable}` placeholder, use with actual values.
- * Template keys follow the same dot-separated pattern as fixed keys.
- *
- * Examples:
- * - Schema: `'scroll.position.${topicId}': number`
- * - Usage: `cacheService.get('scroll.position.topic123')` -> infers `number` type
- *
- * Multiple placeholders are supported:
- * - Schema: `'entity.cache.${type}_${id}': CacheData`
- * - Usage: `cacheService.get('entity.cache.user_456')` -> infers `CacheData` type
+ * Division of labor vs. preferences: `PreferenceService` (SQLite) holds
+ * user-visible configuration; the cache tiers hold recoverable runtime state.
  */
 
 // ============================================================================
 // Template Key Type Utilities
 // ============================================================================
 
-/**
- * Detects whether a key string contains template placeholder syntax.
- *
- * Template keys use `${variable}` syntax to define dynamic segments.
- * This type returns `true` if the key contains at least one `${...}` placeholder.
- *
- * @template K - The key string to check
- * @returns `true` if K contains `${...}`, `false` otherwise
- *
- * @example
- * ```typescript
- * type Test1 = IsTemplateKey<'scroll.position.${id}'>     // true
- * type Test2 = IsTemplateKey<'entity.cache.${a}_${b}'>    // true
- * type Test3 = IsTemplateKey<'app.path.resources'>        // false
- * ```
- */
+/** `true` when the key contains at least one `${...}` placeholder. */
 export type IsTemplateKey<K extends string> = K extends `${string}\${${string}}${string}`
   ? true
   : false;
 
 /**
- * Expands a template key pattern into a matching literal type.
- *
- * Replaces each `${variable}` placeholder with `string`, allowing
- * TypeScript to match concrete keys against the template pattern.
- * Recursively processes multiple placeholders.
- *
- * @template T - The template key pattern to expand
- * @returns A template literal type that matches all valid concrete keys
- *
- * @example
- * ```typescript
- * type Test1 = ExpandTemplateKey<'scroll.position.${id}'>
- * // Result: `scroll.position.${string}` (matches 'scroll.position.123', etc.)
- *
- * type Test2 = ExpandTemplateKey<'entity.cache.${type}_${id}'>
- * // Result: `entity.cache.${string}_${string}` (matches 'entity.cache.user_123', etc.)
- *
- * type Test3 = ExpandTemplateKey<'app.path.resources'>
- * // Result: 'app.path.resources' (unchanged for non-template keys)
- * ```
+ * Expands each `${variable}` placeholder to `string`, so concrete keys
+ * (`scroll.position.topic123`) match their template pattern.
  */
 export type ExpandTemplateKey<T extends string> =
   T extends `${infer Prefix}\${${string}}${infer Suffix}`
     ? `${Prefix}${string}${ExpandTemplateKey<Suffix>}`
     : T;
 
-/**
- * Processes a cache key, expanding template patterns if present.
- *
- * For template keys (containing `${...}`), returns the expanded pattern.
- * For fixed keys, returns the key unchanged.
- *
- * @template K - The key to process
- * @returns The processed key type (expanded if template, unchanged if fixed)
- *
- * @example
- * ```typescript
- * type Test1 = ProcessKey<'scroll.position.${id}'>  // `scroll.position.${string}`
- * type Test2 = ProcessKey<'app.path.resources'>     // 'app.path.resources'
- * ```
- */
+/** Expanded pattern for template keys; the key itself for fixed keys. */
 export type ProcessKey<K extends string> = IsTemplateKey<K> extends true ? ExpandTemplateKey<K> : K;
 
 // ============================================================================
-// Memory Cache Schema
+// Memory Cache Schemas
 // ============================================================================
 
 /**
@@ -135,24 +57,18 @@ export const DefaultUseCache: UseCacheSchema = {
  * observable as `undefined`; entries are not initialized from defaults.
  */
 export type BackendCacheSchema = {
-  // Round-robin cursor for enabled provider API keys. Aligned with Desktop
-  // Main's ProviderService cache key.
+  // Round-robin cursor for enabled provider API keys.
   'settings.provider.${providerId}.last_used_key_id': string;
 };
 
 // ============================================================================
-// Persist Cache Schema
+// Persist Cache Schemas
 // ============================================================================
 
 /**
- * Persist cache schema defining allowed keys and their value types (fixed keys
- * only, no TTL). Values MUST be JSON-serializable (no Date/Map/Set/undefined)
- * and defaults must not be `undefined` — the backing store round-trips every
- * value through JSON.
- *
- * Division of labor vs. preferences: `PreferenceService` (SQLite) holds
- * user-visible configuration; this tier holds recoverable UI state that merely
- * benefits from surviving a restart.
+ * Persist cache schema (fixed keys only, no TTL). Values MUST be
+ * JSON-serializable (no Date/Map/Set/undefined) and defaults must not be
+ * `undefined` — the backing store round-trips every value through JSON.
  */
 export type PersistCacheSchema = {
   // Persist-layer self-test key: exercises the typed persist API and round-trip
@@ -165,8 +81,8 @@ export const DefaultPersistCache: PersistCacheSchema = {
 };
 
 /**
- * Backend-owned persist cache schema. This store is physically independent
- * from the frontend persist cache, matching Desktop Main/Renderer ownership.
+ * Backend-owned persist cache schema, physically independent from the
+ * frontend persist cache.
  */
 export type BackendPersistCacheSchema = {
   'internal.persist_probe': number;
@@ -180,24 +96,13 @@ export const DefaultBackendPersistCache: BackendPersistCacheSchema = {
 // Cache Key Types
 // ============================================================================
 
-/**
- * Key type for persist cache (fixed keys only)
- */
+/** Key type for persist cache (fixed keys only). */
 export type PersistCacheKey = keyof PersistCacheSchema;
 
 /** Backend persist keys are fixed and backend-owned. */
 export type BackendPersistCacheKey = keyof BackendPersistCacheSchema;
 
-/**
- * Key type for memory cache (supports both fixed and template keys).
- *
- * This type expands all schema keys using ProcessKey, which:
- * - Keeps fixed keys unchanged (e.g., 'app.path.resources')
- * - Expands template keys to match patterns (e.g., 'scroll.position.${id}' -> `scroll.position.${string}`)
- *
- * The resulting union type allows TypeScript to accept any concrete key
- * that matches either a fixed key or an expanded template pattern.
- */
+/** Frontend memory keys, including concrete instances of template entries. */
 export type UseCacheKey = {
   [K in keyof UseCacheSchema]: ProcessKey<K & string>;
 }[keyof UseCacheSchema];
@@ -207,21 +112,7 @@ export type BackendCacheKey = {
   [K in keyof BackendCacheSchema]: ProcessKey<K & string>;
 }[keyof BackendCacheSchema];
 
-/**
- * Infers the value type for a given cache key from UseCacheSchema.
- *
- * Works with both fixed keys and template keys:
- * - For fixed keys, returns the exact value type from schema
- * - For template keys, matches the key against expanded patterns and returns the value type
- *
- * If the key doesn't match any schema entry, returns `never`.
- *
- * @example
- * ```typescript
- * type T1 = InferUseCacheValue<'internal.memory_probe.frontend'>  // string
- * type T2 = InferUseCacheValue<'unknown.key'>                                // never
- * ```
- */
+/** Resolves a concrete frontend memory key to its schema value (`never` on miss). */
 export type InferUseCacheValue<K extends string> = {
   [S in keyof UseCacheSchema]: K extends ProcessKey<S & string> ? UseCacheSchema[S] : never;
 }[keyof UseCacheSchema];
@@ -230,20 +121,3 @@ export type InferUseCacheValue<K extends string> = {
 export type InferBackendCacheValue<K extends string> = {
   [S in keyof BackendCacheSchema]: K extends ProcessKey<S & string> ? BackendCacheSchema[S] : never;
 }[keyof BackendCacheSchema];
-
-/**
- * Type guard for casual cache keys that blocks schema-defined keys.
- *
- * Used to ensure casual API methods (getCasual, setCasual, etc.) cannot
- * be called with keys that are defined in the schema (including template patterns).
- * This enforces proper API usage: use type-safe methods for schema keys,
- * use casual methods only for truly dynamic/unknown keys.
- *
- * Known limitation (inherited from desktop): this only bites on literal string
- * arguments — variable keys widen to `string` and pass through unchecked. The
- * runtime `templateToRegex` matching is the real gate.
- *
- * @template K - The key to check
- * @returns `K` if the key doesn't match any schema pattern, `never` if it does
- */
-export type UseCacheCasualKey<K extends string> = K extends UseCacheKey ? never : K;
